testing/web-platform/tests/css/css-text/i18n/css3-text-line-break-baspglwj-128.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>CSS3 Text, linebreaks: U+0F12 TIBETAN MARK RGYA GRAM SHAD</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='https://drafts.csswg.org/css-text-3/#line-break-details'>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name='flags' content='dom'>
<meta name="assert" content="[Exploratory] The browser will NOT break a line of text containing any Unicode character with the GL non-breaking property.">
<style type='text/css'>
#breakable { font-family: monospace; font-size: 25px; width: 50ch; line-height: 30px; }
</style>
</head>
<body>



<div class="test">
			<div id="breakable">aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa&#x0F12;bbbbbbbbbbbbbbbbbbbbbbbbbbbbbb</div>
			</div>


<!--Notes:

Box height: <span id='debugresult'></span><script>document.getElementById('debugresult').innerHTML = document.getElementById('breakable').offsetHeight;</script>

-->
<script>
test(function() {
assert_true(document.getElementById('breakable').offsetHeight < 35);
}, " ");
</script>

<div id='log'></div>

</body>
</html>
